The decimal save pointable was using the wrong type.
git-svn-id: https://svn.apache.org/repos/asf/incubator/vxquery@1369962 13f79535-47bb-0310-9956-ffa450edef68

+++ trunk/vxquery/vxquery-core/src/main/java/org/apache/vxquery/runtime/functions/arithmetic/SubtractOperation.java
@@ -160,7 +160,7 @@ public class SubtractOperation extends AbstractArithmeticOperation {
         // Save
         dOut.write(ValueTag.XS_DECIMAL_TAG);
         dOut.writeByte(place1);
-        dOut.writeDouble(value1);
+        dOut.writeLong(value1);
     }
 
     @Override
